adapt bs4.1.0

+++ src/gausskernel/storage/access/common/reloptions.cpp
@@ -464,7 +464,7 @@ static relopt_string stringRelOpts[] = {
     {
         {"storage_type", "Specifies the Table accessor routines",
          RELOPT_KIND_HEAP | RELOPT_KIND_BTREE | RELOPT_KIND_TOAST},
-        strlen(TABLE_ACCESS_METHOD_ASTORE),
+        (int)strlen(TABLE_ACCESS_METHOD_ASTORE),
         false,
         ValidateStrOptTableAccessMethod,
         TABLE_ACCESS_METHOD_ASTORE,
